Moved all tests to test folder
