+++ shoppinglist_react/src/__tests__/App.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import {render, shallow} from 'enzyme';
 import expect from 'expect';
-import App from "./components/App";
+import App from "../components/App";
 import renderer from 'react-test-renderer';
 import {BrowserRouter} from 'react-router-dom';
 import "./setup_test";
@@ -13,7 +13,7 @@ describe('App component', () => {
 
         const wrapper = render(<BrowserRouter><App/>
         </BrowserRouter>);
-        expect(wrapper.find('div').length).toEqual(8);
+        expect(wrapper.find('div').length).toEqual(9);
 
     });
