+++ shoppinglist_react/src/__tests__/auth.test.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import expect from 'expect';
 import renderer from 'react-test-renderer';
-import Register from "./components/auth/register";
-import Login from "./components/auth/login";
-import SendEmail from "./components/auth/email_send";
-import ResetPassword from "./components/auth/reset_password";
-import Header from "./components/header";
+import Register from "../components/auth/register";
+import Login from "../components/auth/login";
+import SendEmail from "../components/auth/email_send";
+import ResetPassword from "../components/auth/reset_password";
+import Header from "../components/header";
 import "./setup_test";
