+++ shoppinglist_react/src/__tests__/email_send.test.js
@@ -2,7 +2,7 @@ import moxios from 'moxios';
 import {mount, shallow} from 'enzyme';
 import React from 'react';
 import "./setup_test";
-import SendEmail from './components/auth/email_send'
+import SendEmail from '../components/auth/email_send'
 
 
 describe('mocking email send requests', function () {
