+++ shoppinglist_react/src/__tests__/items.test.js
@@ -2,9 +2,9 @@ import React from 'react';
 import expect from 'expect';
 import renderer from 'react-test-renderer';
 import "./setup_test"
-import Items from "./components/items/items";
-import AddItem from "./components/items/additem";
-import ItemContents from "./components/items/itemcontents";
+import Items from "../components/items/items";
+import AddItem from "../components/items/additem";
+import ItemContents from "../components/items/itemcontents";
 
 
 describe('Shopping list components', () => {
