+++ shoppinglist_react/src/__tests__/lists.test.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import expect from 'expect';
 import renderer from 'react-test-renderer';
-import ShoppingList from "./components/shoppinglist/shoppinglist";
-import AddList from "./components/shoppinglist/addlist";
-import TableContents from "./components/shoppinglist/tablecontents";
+import ShoppingList from "../components/shoppinglist/shoppinglist";
+import AddList from "../components/shoppinglist/addlist";
+import TableContents from "../components/shoppinglist/tablecontents";
 import "./setup_test"
