+++ shoppinglist_react/src/__tests__/setup_test.js
@@ -1,3 +1,4 @@
+// import React from 'react';
 import Adapter from 'enzyme-adapter-react-16';
 import {configure} from 'enzyme';
